requirements.txt
streamlit==1.51.0
langchain==1.0.5
langchain-community==0.4.1
langchain-pinecone==0.2.13
langchain-huggingface==1.0.1
langchain-text-splitters==1.0.0
pinecone-client==2.4.0
tqdm==4.67.0
huggingface_hub==0.36.1
python-dotenv==1.2.0
pypdf==6.2.0

datasets==4.4.1
scikit-learn==1.7.2
numpy==2.3.4
pandas==2.3.3
